+++ src/binary/mockedRunProcess.ts
@@ -5,7 +5,7 @@ import * as stream from "stream";
 import { anyFunction, anyString, instance, mock, when } from "ts-mockito";
 import { BinaryProcessRun } from "./runProcess";
 import { isFunction } from "../utils/utils";
-import { Capability } from "../capabilities";
+import { Capability } from "../capabilities/capabilities";
 import { BinaryGenericRequest } from "../test/suite/utils/helper";
 
 export const spawnedProcessMock: child_process.ChildProcess = mock<child_process.ChildProcess>();
