+++ src/capabilities/getSuggestionMode.ts
