+++ src/inlineSuggestions/textListener.ts
@@ -3,7 +3,7 @@ import {
   getCurrentSuggestion,
   setSuggestionsState,
 } from "./inlineSuggestionState";
-import runCompletion from "./runCompletion";
+import runCompletion from "../runCompletion";
 import setInlineSuggestion from "./setInlineSuggestion";
 
 export default async function textListener(
