+++ src/validator/ValidatorClient.ts
@@ -4,7 +4,7 @@ import * as child_process from "child_process";
 import * as readline from "readline";
 import * as vscode from "vscode";
 import setState from "../binary/requests/setState";
-import { Capability, isCapabilityEnabled } from "../capabilities";
+import { Capability, isCapabilityEnabled } from "../capabilities/capabilities";
 import { StatePayload } from "../globals/consts";
 import CancellationToken from "./CancellationToken";
 import {
